From 899314c0a45c54808b26b379fe333dd409d26fd1 Mon Sep 17 00:00:00 2001 From: Venkat Devarasetty <vdevaras@codeaurora.org> Date: Mon, 25 Mar 2013 23:37:45 +0530 Subject: [PATCH 0649/6902] msm: pm: avoid array overrun for msm_pm_sleep_modes
The array is not defined for the maximum possible size.
So it is possible that the array may overrun when it is
dereferenced using MSM_PM_MODE macro.

Add a dummy entry to make sure array is of max size.

Change-Id: I5783b0aa6c8295c5c5aabcb498700c6af1a3eba5
Signed-off-by: Venkat Devarasetty <vdevaras@codeaurora.org>
---
 arch/arm/mach-msm/pm-data.c | 7 +++++++
 1 file changed, 7 insertions(+)

diff --git a/arch/arm/mach-msm/pm-data.c b/arch/arm/mach-msm/pm-data.c
index 7fe7795..7df9855 100644
--- a/arch/arm/mach-msm/pm-data.c
+++ b/arch/arm/mach-msm/pm-data.c
@@ -125,4 +125,11 @@ struct msm_pm_platform_data msm_pm_sleep_modes[] = {
 		.idle_enabled = 1,
 		.suspend_enabled = 0,
 	},
+
+	[MSM_PM_MODE(3, MSM_PM_SLEEP_MODE_NR)] = {
+		.idle_supported = 0,
+		.suspend_supported = 0,
+		.idle_enabled = 0,
+		.suspend_enabled = 0,
+	},
 };
--
2.8.1